requirements.txt
caldav==1.3.9
geojson==3.1.0
icalendar==6.0.0
Jinja2==3.1.4
nicegui==2.3.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests~=2.28.2
typing_extensions==4.12.2
urllib3~=1.26.15
python-telegram-bot~=21.6